requirements.txt
ultralytics
opencv-python

tqdm
pillow